- removed old plot editor iterations - tidied up widget creator
