+++ Frontend/package.json
@@ -1,7 +1,7 @@
 {
   "name": "fcc-dashboard-alpha",
   "version": "1.0.0",
-  "description": "Alpha version of the Future Cities Catapult Sharing Cities Dashboard",
+  "description": "Future Cities Catapult Sharing Cities Dashboard",
   "main": "index.js",
   "scripts": {
     "start": "webpack-dev-server --mode development --open --hot",
@@ -21,7 +21,6 @@
     "axios": "^0.18.0",
     "classnames": "^2.2.6",
     "color": "^3.1.0",
-    "jiminy": "^0.1.1",
     "jwt-decode": "^2.2.0",
     "leaflet": "^1.3.4",
     "lodash": "^4.17.11",
@@ -74,6 +73,6 @@
     "vega": "^4.4.0",
     "webpack": "^4.24.0",
     "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.10"
+    "webpack-dev-server": "^3.1.14"
   }
 }
