+++ Frontend/src/actions/__tests__/editorActions.spec.js
@@ -2,18 +2,17 @@ import configureStore from 'redux-mock-store';
 import * as editorActions from './../editorActions';
 import {
   PURGE_EDITOR,
+  SET_WIDGET_NAME,
+  SET_WIDGET_TYPE,
+  SET_PLOT_DATA,
+  SET_PLOT_PROPERTY,
+  SET_PLOT_ENCODING,
   SET_MAP_CENTER,
   SET_MAP_HEATMAP_ATTRIBUTE,
   SET_MAP_IS_MAPPABLE,
   SET_MAP_SHOW_HEATMAP,
   SET_MAP_TILE_LAYER,
   SET_MAP_ZOOM,
-  SET_PLOT_DATA,
-  SET_PLOT_DESCRIPTION,
-  SET_PLOT_ENCODING,
-  SET_PLOT_TYPE,
-  SET_WIDGET_NAME,
-  SET_WIDGET_TYPE,
 } from "./../../constants";
 
 const mockStore = configureStore();
@@ -49,7 +48,98 @@ describe('editorActions', () => {
 
       store.dispatch(editorActions.purgeEditor());
 
-      expect(store.getActions()).toEqual(expectedActions);
+      expect(store.getActions()).toEqual(expectedActions)
+    })
+  });
+
+  describe('setWidgetName', () => {
+    test('Dispatches the correct action and payload', () => {
+      let expectedActions = [
+        {
+          type: SET_WIDGET_NAME,
+          payload: 'Bike Availability',
+        },
+      ];
+
+      store.dispatch(editorActions.setWidgetName('Bike Availability'));
+
+      expect(store.getActions()).toEqual(expectedActions)
+    })
+  });
+
+  describe('setWidgetType', () => {
+    test('Dispatches the correct action and payload', () => {
+      let expectedActions = [
+        {
+          type: SET_WIDGET_TYPE,
+          payload: 'plot',
+        },
+      ];
+
+      store.dispatch(editorActions.setWidgetType('plot'));
+
+      expect(store.getActions()).toEqual(expectedActions)
+    })
+  });
+
+  describe('setPlotData', () => {
+    test('Dispatches the correct action and payload', () => {
+      let expectedActions = [
+        {
+          type: SET_PLOT_DATA,
+          payload: [{a: 1}],
+        },
+      ];
+
+      store.dispatch(editorActions.setPlotData([{ a: 1 }]));
+
+      expect(store.getActions()).toEqual(expectedActions)
+    })
+  });
+
+  describe('setPlotProperty', () => {
+    test('Dispatches the correct action and payload', () => {
+      let expectedActions = [
+        {
+          type: SET_PLOT_PROPERTY,
+          payload: {
+            property: "mark",
+            value: "point",
+          },
+        },
+      ];
+
+      store.dispatch(editorActions.setPlotProperty({
+        property: "mark",
+        value: "point",
+      }));
+
+      expect(store.getActions()).toEqual(expectedActions)
+    })
+  });
+
+  describe('setPlotEncoding', () => {
+    test('Dispatches the correct action and payload', () => {
+      let expectedActions = [
+        {
+          type: SET_PLOT_ENCODING,
+          payload: {
+            x: {
+              field: "Timestamp",
+              type: "temporal",
+            },
+          },
+        },
+      ];
+
+      store.dispatch(editorActions.setPlotEncoding({
+        x: {
+          field: "Timestamp",
+          type: "temporal",
+        },
+      }));
+
+      expect(store.getActions()).toEqual(expectedActions)
     })
   });
 
@@ -67,7 +157,7 @@ describe('editorActions', () => {
 
       store.dispatch(editorActions.setMapCenter({ lat: 51.505, lng: -0.09 }));
 
-      expect(store.getActions()).toEqual(expectedActions);
+      expect(store.getActions()).toEqual(expectedActions)
     })
   });
 
@@ -82,7 +172,7 @@ describe('editorActions', () => {
 
       store.dispatch(editorActions.setMapHeatmapAttribute('bikes'));
 
-      expect(store.getActions()).toEqual(expectedActions);
+      expect(store.getActions()).toEqual(expectedActions)
     })
   });
 
@@ -97,7 +187,7 @@ describe('editorActions', () => {
 
       store.dispatch(editorActions.setMapIsMappable(true));
 
-      expect(store.getActions()).toEqual(expectedActions);
+      expect(store.getActions()).toEqual(expectedActions)
     })
   });
 
@@ -112,7 +202,7 @@ describe('editorActions', () => {
 
       store.dispatch(editorActions.setMapShowHeatmap(true));
 
-      expect(store.getActions()).toEqual(expectedActions);
+      expect(store.getActions()).toEqual(expectedActions)
     })
   });
 
@@ -127,7 +217,7 @@ describe('editorActions', () => {
 
       store.dispatch(editorActions.setMapTileLayer('CartoDB.DarkMatter'));
 
-      expect(store.getActions()).toEqual(expectedActions);
+      expect(store.getActions()).toEqual(expectedActions)
     })
   });
 
@@ -142,105 +232,7 @@ describe('editorActions', () => {
 
       store.dispatch(editorActions.setMapZoom(13));
 
-      expect(store.getActions()).toEqual(expectedActions);
+      expect(store.getActions()).toEqual(expectedActions)
     })
   });
-
-  describe('setPlotData', () => {
-    test('Dispatches the correct action and payload', () => {
-      let expectedActions = [
-        {
-          type: SET_PLOT_DATA,
-          payload: [{a: 1}],
-        },
-      ];
-
-      store.dispatch(editorActions.setPlotData([{ a: 1 }]));
-
-      expect(store.getActions()).toEqual(expectedActions);
-    })
-  });
-
-  describe('setPlotDescription', () => {
-    test('Dispatches the correct action and payload', () => {
-      let expectedActions = [
-        {
-          type: SET_PLOT_DESCRIPTION,
-          payload: 'Availability of bikes',
-        },
-      ];
-
-      store.dispatch(editorActions.setPlotDescription('Availability of bikes'));
-
-      expect(store.getActions()).toEqual(expectedActions);
-    })
-  });
-
-  describe('setPlotEncoding', () => {
-    test('Dispatches the correct action and payload', () => {
-      let expectedActions = [
-        {
-          type: SET_PLOT_ENCODING,
-          payload: {
-            axis: 'x',
-            field: 'bikes',
-            type: 'quantitative',
-          },
-        },
-      ];
-
-      store.dispatch(editorActions.setPlotEncoding({
-        axis: 'x',
-        field: 'bikes',
-        type: 'quantitative',
-      }));
-
-      expect(store.getActions()).toEqual(expectedActions);
-    })
-  });
-
-  describe('setPlotType', () => {
-    test('Dispatches the correct action and payload', () => {
-      let expectedActions = [
-        {
-          type: SET_PLOT_TYPE,
-          payload: 'bar',
-        },
-      ];
-
-      store.dispatch(editorActions.setPlotType('bar'));
-
-      expect(store.getActions()).toEqual(expectedActions);
-    })
-  });
-
-  describe('setWidgetName', () => {
-    test('Dispatches the correct action and payload', () => {
-      let expectedActions = [
-        {
-          type: SET_WIDGET_NAME,
-          payload: 'Bike Availability',
-        },
-      ];
-
-      store.dispatch(editorActions.setWidgetName('Bike Availability'));
-
-      expect(store.getActions()).toEqual(expectedActions);
-    })
-  });
-
-  describe('setWidgetType', () => {
-    test('Dispatches the correct action and payload', () => {
-      let expectedActions = [
-        {
-          type: SET_WIDGET_TYPE,
-          payload: 'plot',
-        },
-      ];
-
-      store.dispatch(editorActions.setWidgetType('plot'));
-
-      expect(store.getActions()).toEqual(expectedActions);
-    })
-  })
 });
