--- Frontend/src/actions/vegaLiteActions.js
@@ -1,4 +1,0 @@
-import {
-  VEGA_LITE_SET_PROPERTY,
-  VEGA_LITE_SET_ENCODING_CHANNEL,
-} from './../constants';
