--- Frontend/src/components/vega/LayerDialog.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const styles = theme => ({
-
-});
-
-class LayerDialog extends React.Component {
-
-}
-
-export default LayerDialog;
